Add fast-path for binding security check of DOMWindow https://bugs.webkit.org/show_bug.cgi?id=226930
Reviewed by Geoffrey Garen.

The security check[1] must pass if the current JSDOMGlobalObject is the same to the accessed JSDOMWindow.
This clarification paves the way to emit JIT code which removes this security check when the lexical and
accessed JSGlobalObjects are the same.

[1]: https://html.spec.whatwg.org/multipage/browsers.html#integration-with-idl

* bindings/js/JSDOMBindingSecurity.cpp:
(WebCore::BindingSecurity::shouldAllowAccessToDOMWindow):
* bindings/js/JSDOMBindingSecurity.h:
* bindings/scripts/CodeGeneratorJS.pm:
(GenerateAttributeGetterBodyDefinition):
(GenerateAttributeSetterBodyDefinition):
(GenerateOperationBodyDefinition):

Canonical link: https://commits.webkit.org/238739@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@278782 268f45cc-cd09-0410-ab3c-d52691b4dbfc