SimulatedXRDevice::shutDownTrackingAndRendering() should clear it's GraphicsContextGL to ensure the resource is cleaned up quickly https://bugs.webkit.org/show_bug.cgi?id=226947
Reviewed by Dean Jackson.

SimulatedXRDevice::shutDownTrackingAndRendering() needs to clear its GraphicsContextGL
to ensure the resource is cleaned up quickly, and doesn't wait for GC to be reclaimed.
Without this, we can run out of GraphicsContextGLs and fail tests if GC happens to be
running a bit behind.

* testing/WebFakeXRDevice.cpp:
(WebCore::SimulatedXRDevice::shutDownTrackingAndRendering):


Canonical link: https://commits.webkit.org/238773@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@278820 268f45cc-cd09-0410-ab3c-d52691b4dbfc