Don't look at the (non-existent) child2 of DelById https://bugs.webkit.org/show_bug.cgi?id=227095
Reviewed by Mark Lam.

Trivial fix to a broken rebase: while it is ok to share most code between DelById and DelByVal, only the latter has a child2(), so it should not be accessed if we are compiling the former.
No new test, as it was caught by one of our existing tests.

* ftl/FTLLowerDFGToB3.cpp:
(JSC::FTL::DFG::LowerDFGToB3::compileDelBy):


Canonical link: https://commits.webkit.org/238887@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@278959 268f45cc-cd09-0410-ab3c-d52691b4dbfc