Fix speculated type in the one-argument overload of speculateNeitherDoubleNorHeapBigIntNorString https://bugs.webkit.org/show_bug.cgi?id=227119
Reviewed by Yusuke Suzuki.

Same problem as bug 226786: a missing check for HeapBigInt in the speculateNeitherDoubleNorHeapBigIntNorString function introduced in 226676.

I also rewrote the SpeculatedType for NeitherDoubleNorHeapBigIntNorString in typeFilterFor for readability. The old and the new SpeculatedType are perfectly identical, it is just a different (and in my view more readable) way of writing it.

* dfg/DFGSpeculativeJIT.cpp:
(JSC::DFG::SpeculativeJIT::speculateNeitherDoubleNorHeapBigIntNorString):
* dfg/DFGUseKind.h:
(JSC::DFG::typeFilterFor):


Canonical link: https://commits.webkit.org/238972@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@279052 268f45cc-cd09-0410-ab3c-d52691b4dbfc