[JSC] Fix consistency check during stack splitting in Wasm::LLIntGenerator::addLoop https://bugs.webkit.org/show_bug.cgi?id=226012
Patch by Xan Lopez <xan@igalia.com> on 2021-06-21
Reviewed by Tadeu Zagallo.

JSTests:

* stress/wasm-loop-consistency.js: Added.
(vm.isWasmSupported):

Source/JavaScriptCore:

It is possible for the wasm llint generator to call
checkConsistency() on a stack that is only halfway through being
properly setup. Specifically, when generating a loop block, we use
splitStack() to pop the arguments for the loop into a new stack,
and materializeConstantsAndLocals() to materialize the constants
and aliases in the loop arguments, but the arguments won't be
added back to the stack until the very end of the loop code
generation. Since materializeConstantsAndLocals() will check the
correctness of the expression stack, which isn't yet fully formed,
we'll fail its ASSERT. To workaround this, we create a variant of
materializeConstantsAndLocals() that does not check for
correctness (similar to what we do in push()), and manually check
the correctness of the new split stack in
Wasm::LLIntGenerator::addLoop(), which is the place that knows the
details of this intermediate state.

For more details, see: https://bugs.webkit.org/show_bug.cgi?id=226012#c8

* wasm/WasmLLIntGenerator.cpp:
(JSC::Wasm::LLIntGenerator::checkConsistencyOfExpressionStack):
(JSC::Wasm::LLIntGenerator::checkConsistency):
(JSC::Wasm::LLIntGenerator::materializeConstantsAndLocals):
(JSC::Wasm::LLIntGenerator::addLoop):

Canonical link: https://commits.webkit.org/239001@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@279082 268f45cc-cd09-0410-ab3c-d52691b4dbfc