[iOS 15] Fix the internal build after rdar://76549109 https://bugs.webkit.org/show_bug.cgi?id=227430 rdar://79832425
Reviewed by Sam Weinig.

Fix the build by replacing the use of the `UIFloatIsZero` macro with its actual definition. The immediate cause
of this build failure is a UIKit change that changed the definition of `UIFloatIsZero`, such that we end up
redefining `UIFloatIsZero` in HTMLConverter.mm. However, since `UIFloatIsZero` is specific to iOS family and
HTMLConverter is used only in this one place on both macOS and iOS, it makes more sense to simply use the macro
definition itself.

* editing/cocoa/HTMLConverter.mm:
(HTMLConverter::computedAttributesForElement):


Canonical link: https://commits.webkit.org/239196@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@279320 268f45cc-cd09-0410-ab3c-d52691b4dbfc