ANGLE Metal primitive restart range computation should not be done unless primitive restart is enabled https://bugs.webkit.org/show_bug.cgi?id=227452
Patch by Kimmo Kinnunen <kkinnunen@apple.com> on 2021-06-29
Reviewed by Kenneth Russell.

Primitive restart needs to scan the whole index buffer for values.
This should not be done unless needed.

* src/libANGLE/renderer/metal/VertexArrayMtl.mm:
(rx::VertexArrayMtl::getDrawIndices):
First check for primitive restart enabledness.
Compute the primitive restart values after this.
The algoritm should succeed when the count of primitive
restart ranges is zero, and it does not do much of work.
Use the restart range vector as reference, so that no copy
would be done.
Add spaces the if statements, comments and references.
Move the else arm of the if statement with return statement
to body of the function to reduce nesting.

Canonical link: https://commits.webkit.org/239238@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@279372 268f45cc-cd09-0410-ab3c-d52691b4dbfc