RealtimeIncomingAudioSourceCocoa should support other sample rate than 48000 https://bugs.webkit.org/show_bug.cgi?id=227439
Reviewed by Eric Carlson.

Source/WebCore:

We reduced memory allocations by early allocating buffers for the most often used sample rate (48000), but this broke other sample rates, like for G722.
Make sure to update sample rate and buffers if it is not 48000.
Since the initial OnData call is often with a sample rate of 16000 and switches to 48000 after a few calls, early return for a few initial calls (20) or if sample rate is not 16000.
Covered by updated test.

* platform/mediastream/mac/RealtimeIncomingAudioSourceCocoa.cpp:
(WebCore::RealtimeIncomingAudioSourceCocoa::OnData):
* platform/mediastream/mac/RealtimeIncomingAudioSourceCocoa.h:

LayoutTests:

Improve test to validate received audio is not silence.

* webrtc/audio-peer-connection-g722-expected.txt:
* webrtc/audio-peer-connection-g722.html:


Canonical link: https://commits.webkit.org/239315@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@279459 268f45cc-cd09-0410-ab3c-d52691b4dbfc