URL path setter should add "./" before non-empty path and after empty host for URLs with non-special schemes https://bugs.webkit.org/show_bug.cgi?id=227750
Reviewed by Chris Dumez.

LayoutTests/imported/w3c:

* web-platform-tests/url/url-setters-a-area.window-expected.txt:
* web-platform-tests/url/url-setters.any-expected.txt:
* web-platform-tests/url/url-setters.any.worker-expected.txt:

Source/WTF:

This is related to r267837 but in the path setter, too.
This helps preserve idempotence in parsed URLs.
This is an edge case that will rarely be hit in practice, but now the path setter matches the parser.

* wtf/URL.cpp:
(WTF::URL::setPath):



Canonical link: https://commits.webkit.org/239475@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@279662 268f45cc-cd09-0410-ab3c-d52691b4dbfc