Source/WebCore: [GPU Process] Draw PDFs using an intermediate ImageBuffer when using the GPU process https://bugs.webkit.org/show_bug.cgi?id=227550
Reviewed by Tim Horton.

When drawing PDFs, we usually already have a cached ImageBuffer of the contents of the PDF.
However, in the situation where we don't have this cached ImageBuffer, we can't just draw
the PDF into the platformContext when we're using the GPU Process. So, instead, we need to
draw the PDF into an intermediate ImageBuffer and then send that to the GPU Process.

When we have a fully functional custom CGContext, we can use that to send all the PDF drawing
commands to the GPU Process, and this patch should then be reverted.

Test: fast/images/pdf-in-canvas.html

* platform/graphics/cg/PDFDocumentImage.cpp:
(WebCore::PDFDocumentImage::draw):
* platform/mac/LocalCurrentGraphicsContextMac.mm:
(WebCore::LocalCurrentGraphicsContext::LocalCurrentGraphicsContext):

LayoutTests:
[GPU Process] Temporarily disable drawing large PDFs in display list drawing
https://bugs.webkit.org/show_bug.cgi?id=227550

Reviewed by Tim Horton.

* fast/images/pdf-in-canvas-expected.html: Added.
* fast/images/pdf-in-canvas.html: Added.

Canonical link: https://commits.webkit.org/239522@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@279741 268f45cc-cd09-0410-ab3c-d52691b4dbfc