[WebCrypto] deriveBits() incorrectly throws for PBKDF2 algorithm when the password is empty https://bugs.webkit.org/show_bug.cgi?id=227810
Reviewed by Yusuke Suzuki.

LayoutTests/imported/w3c:

Rebaseline WPT tests that are now passing.

* web-platform-tests/WebCryptoAPI/derive_bits_keys/pbkdf2.https.any-expected.txt:
* web-platform-tests/WebCryptoAPI/derive_bits_keys/pbkdf2.https.any.worker-expected.txt:

Source/WebCore:

deriveBits() was incorrectly throwing for PBKDF2 algorithm when the provided password was empty.
The reason for that was that we would pass a null key pointer to CCKeyDerivationPBKDF() in this
case and it was causing CCKeyDerivationPBKDF() to return an error, even though the key length is
0. To avoid this issue, we now pass an empty string as key to CCKeyDerivationPBKDF() instead of
a null pointer, when the length is 0.

This is causing Safari to fail subtests in the following WPT test:
- https://wpt.live/WebCryptoAPI/derive_bits_keys/pbkdf2.https.any.html?5001-6000

       Empty passwords are supported by both Chrome and Firefox.

No new tests, rebaselined existing tests.

* crypto/mac/CryptoAlgorithmPBKDF2Mac.cpp:
(WebCore::CryptoAlgorithmPBKDF2::platformDeriveBits):


Canonical link: https://commits.webkit.org/239555@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@279788 268f45cc-cd09-0410-ab3c-d52691b4dbfc