Fix a leak of BBitmap instances
The main problem was a missing adoptRef in ImageBufferDataHaiku
constructor, which means the bitmaps would have an extra reference and
never be destroyed.

While I'm at it, make the ImageBufferData class not use NativeImage as
an indirection, and instead use a BBitmap (PlatformImagePtr) directly.
This makes it able to draw only on GraphicsContextHaiku but I don't
think we use any other type of graphics contexts currently.