Regression(r279601) ProcessAssertion may get destroyed on a background thread https://bugs.webkit.org/show_bug.cgi?id=227875 <rdar://76972252>
Reviewed by Geoffrey Garen.

r279601 added an internal WorkQueue to ProcessAssertion, so that we could acquire the RunningBoard assertion
asynchronously on the background queue. When dispatching to the background queue, we capture |protectedThis|,
which means that ProcessAssertion may now get destroyed on the background queue. ProcessAssertion is a main
thread object and destroying it on a non-main thread can lead to crashes.

* UIProcess/ProcessAssertion.h:
* UIProcess/ios/ProcessAssertionIOS.mm:
(WebKit::ProcessAssertion::ProcessAssertion):
(WebKit::ProcessAssertion::~ProcessAssertion):


Canonical link: https://commits.webkit.org/239595@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@279835 268f45cc-cd09-0410-ab3c-d52691b4dbfc